requirements2.txt
absl-py==2.1.0
accelerate==0.28.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
bcrypt==4.1.2
beautifulsoup4==4.12.3
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
cloudinary==1.39.0
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
deep-translator==1.11.4
dnspython==2.6.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
etils==1.8.0
fastapi==0.110.0
filelock==3.13.1
flatbuffers==24.3.7
flax==0.8.2
fonttools==4.49.0
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
googletrans==3.1.0a0
grpcio==1.62.1
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.21.4
hyperframe==5.2.0
idna==2.10
importlib_resources==6.3.2
jax==0.4.25
jaxlib==0.4.25
Jinja2==3.1.3
joblib==1.3.2
keras==3.1.1
kiwisolver==1.4.5
langcodes==3.3.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
namex==0.0.7
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
opt-einsum==3.3.0
optax==0.2.1
optree==0.10.0
orbax-checkpoint==0.5.6
packaging==24.0
pandas==2.2.1
passlib==1.7.4
pillow==10.2.0
plotly==5.20.0
preshed==3.0.9
protobuf==4.25.3
psutil==5.9.8
pusher==3.3.2
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
pymongo==4.6.2
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rich==13.7.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentencepiece==0.2.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.16.1
tensorflow-metal==1.1.0
tensorstore==0.1.56
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
toolz==0.12.1
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.18.1
